# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.Album:
    properties:
      artist:
        type: string
      id:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  dto.HTTPError:
    properties:
      message:
        example: something went wrong, bad request, not_found, unathorized ...
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: get awailable methods
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Album'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Show an index page
      tags:
      - index
  /albums:
    get:
      consumes:
      - application/json
      description: getAlbums responds with the list of all albums as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Album'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: getAlbums responds with the list of all albums as JSON.
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: add by json account
      parameters:
      - description: Add account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.Album'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Album'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Add an album
      tags:
      - albums
  /albums/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Album'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Show an album
      tags:
      - albums
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"